requirements.txt
# requirements.txt

numpy
pandas
scikit-learn
joblib
xgboost
optuna
rdkit
mordred
requests
PyYAML
tqdm

# ---------------------------------------------------------------------
# ПРИМЕЧАНИЕ О СЛОЖНЫХ И ВНЕШНИХ ЗАВИСИМОСТЯХ:
#
# REINVENT:
#    Этот проект использует REINVENT как внешний инструмент для генерации
#    молекул. REINVENT не устанавливается через pip вместе с этими
#    зависимостями. Его необходимо установить отдельно согласно
#    официальной инструкции: https://github.com/MolecularAI/REINVENT4
#    После установки убедитесь, что команда 'reinvent' доступна
#    в вашем системном PATH.
# ---------------------------------------------------------------------
